api_yamdb/requirements.txt
asgiref==3.6.0
atomicwrites==1.4.1
attrs==22.2.0
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
Django==3.2.10
django-adaptors==0.2.5
django-filter==22.1
django-nose==1.4.5
django-shortcuts==1.6
django_debug_toolbar==3.8.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
gunicorn==20.0.4
idna==3.4
importlib-metadata==4.2.0
iniconfig==2.0.0
isort==5.11.4
mypy-extensions==0.4.3
nose==1.2.1
packaging==23.0
pathspec==0.10.3
platformdirs==2.6.2
pluggy==0.13.1
psycopg2-binary==2.9.5
py==1.11.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.14
zipp==3.11.0
